% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{filter_cyberframe_times}
\alias{filter_cyberframe_times}
\title{Returns cyberframe object with only the data from a specific time range}
\usage{
filter_cyberframe_times(obj, times)
}
\arguments{
\item{obj}{Cyberframe object}

\item{times}{Vector of length 2 with start and end times}
}
\value{
Cyberframe object with only the data from the specified time range
}
\description{
Returns cyberframe object with only the data from a specific time range
}
