% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{filter_trial}
\alias{filter_trial}
\title{Returns cyberframe object with only the data from a single trial}
\usage{
filter_trial(obj, index)
}
\arguments{
\item{obj}{Cyberframe object}

\item{index}{Index of the trial. Zero based}
}
\value{
Cyberframe object with only the data from a single trial
}
\description{
Returns cyberframe object with only the data from a single trial
}
