% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_experiment_log}
\alias{get_experiment_log}
\title{Returns the experiment log from the cyberframe object}
\usage{
get_experiment_log(obj)
}
\arguments{
\item{obj}{Cyberframe object}
}
\value{
list with experiment log data
}
\description{
Returns the experiment log from the cyberframe object
}
