% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_position}
\alias{get_position}
\title{Returns the position data from the cyberframe object}
\usage{
get_position(obj)
}
\arguments{
\item{obj}{Cyberframe object}
}
\value{
navr::object with position data
}
\description{
Returns the position data from the cyberframe object
}
