% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_trial_position}
\alias{get_trial_position}
\title{Returns the position data from the cyberframe object for a single trial}
\usage{
get_trial_position(obj, iTrial)
}
\arguments{
\item{obj}{Cyberframe object}

\item{iTrial}{Index of the trial}
}
\value{
navr::object with position data
}
\description{
Returns the position data from the cyberframe object for a single trial
}
