% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{load_cyberframe_log}
\alias{load_cyberframe_log}
\title{Loads a generic cyberframe framework log. These logs have specific header and
data notations. The default logs should be saved in UTF-8 encoding with semicolon 
as a separator, dot for decimal points and unquoted strings. 
The function loads the header and the data frame.}
\usage{
load_cyberframe_log(filepath, func = NULL, ...)
}
\arguments{
\item{filepath}{path to the log}

\item{func}{optional loading function, it loads the log instead of the
default}

\item{...}{additional parameters passed to read.table. skip, sep, header,
stringsAsFactos and encoding cannot be changed}
}
\value{
list with parsed data and optionally $data field with log's dataframe
}
\description{
Loads a generic cyberframe framework log. These logs have specific header and
data notations. The default logs should be saved in UTF-8 encoding with semicolon 
as a separator, dot for decimal points and unquoted strings. 
The function loads the header and the data frame.
}
