% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading.R
\name{open_experiment_logs}
\alias{open_experiment_logs}
\title{Iterates over all __experiment_ files in a folder an saves
them one by one to a return list}
\usage{
open_experiment_logs(dir, exp_timestamp = NULL, flatten = FALSE)
}
\arguments{
\item{exp_timestamp}{time of the}

\item{flatten}{in case of only a single list is returned, unnests the list.
Beware, unnested list causes issues with opening experiments}

\item{directory}{directory where the file is located}
}
\description{
Iterates over all __experiment_ files in a folder an saves
them one by one to a return list
}
