% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{plot_trial_path}
\alias{plot_trial_path}
\alias{plot_trial_path.cyberframe}
\title{Plot the path of a trial}
\usage{
plot_trial_path(obj, iTrial, ...)

\method{plot_trial_path}{cyberframe}(obj, iTrial, ...)
}
\arguments{
\item{obj}{Cyberframe object}

\item{iTrial}{Index of the trial}

\item{...}{Additional arguments passed to navr::plot_path}
}
\description{
Plot the path of a trial
}
\section{Methods (by class)}{
\itemize{
\item \code{plot_trial_path(cyberframe)}: Plot the path of a trial

}}
