% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarise_trials}
\alias{summarise_trials}
\title{Makes a quick summary of the given trials}
\usage{
summarise_trials(obj, indices = c())
}
\arguments{
\item{obj}{Cyberframe object}

\item{index}{indices of the trials to summarise. In the original
format, which is generally zero based.}
}
\value{
Dataframe with summary of the trials
}
\description{
Makes a quick summary of the given trials
}
